# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Brothers App following Clean Architecture.",
        "title": "Brothers App API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticates an active user using a username or email address and password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in with username or email",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes the current authenticated session.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log out",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.EmptyResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/otp/send": {
            "post": {
                "description": "Sends an OTP for registration or password reset. Password reset accepts either email or username and returns the recipient email when delivery is started.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Send OTP",
                "parameters": [
                    {
                        "description": "OTP purpose and email or username",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.SendOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.StartResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/password/reset": {
            "post": {
                "description": "Replaces the password using a reset-only token and revokes the active session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Reset password",
                "parameters": [
                    {
                        "description": "Reset token and new password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/password/verify": {
            "post": {
                "description": "Exchanges an email-delivered reset OTP for a single-use reset token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify password-reset OTP",
                "parameters": [
                    {
                        "description": "Email and OTP",
                        "name": "verification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.OTPVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.ResetVerifyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Rotates the opaque refresh token and returns a new token pair.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh session tokens",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Verifies the registration OTP, creates the user, and issues a token pair. Request an OTP first through POST /auth/otp/send.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "Registration payload",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns safe profile fields for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get current profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates only provided names, avatar URL, and language fields.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Update current profile",
                "parameters": [
                    {
                        "description": "Editable profile fields",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/authdto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "authdto.AuthData": {
            "type": "object",
            "properties": {
                "tokens": {
                    "$ref": "#/definitions/authdto.TokenData"
                },
                "user": {
                    "$ref": "#/definitions/authdto.UserData"
                }
            }
        },
        "authdto.AuthResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/authdto.AuthData"
                },
                "message": {
                    "type": "string",
                    "example": "Request processed successfully"
                },
                "meta": {
                    "$ref": "#/definitions/authdto.ResponseMeta"
                },
                "slug": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "authdto.EmptyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "x-nullable": true
                },
                "message": {
                    "type": "string",
                    "example": "Request processed successfully"
                },
                "meta": {
                    "$ref": "#/definitions/authdto.ResponseMeta"
                },
                "slug": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "authdto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1400
                },
                "data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/responses.ErrorDetails"
                        }
                    ],
                    "x-nullable": true
                },
                "message": {
                    "type": "string",
                    "example": "Некорректные данные"
                },
                "meta": {
                    "$ref": "#/definitions/authdto.ResponseMeta"
                },
                "slug": {
                    "type": "string",
                    "example": "invalid_data"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "authdto.LoginRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "qobulov"
                },
                "password": {
                    "type": "string",
                    "example": "strong-password"
                }
            }
        },
        "authdto.OTPVerifyRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "ali@example.com"
                },
                "otp": {
                    "type": "string",
                    "example": "482910"
                }
            }
        },
        "authdto.RefreshRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "opaque-refresh-token"
                }
            }
        },
...
        "authdto.RegisterRequest": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "email": {
                    "type": "string",
                    "example": "ali@example.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "Azizbek"
                },
                "language": {
                    "type": "string",
                    "example": "uz"
                },
                "last_name": {
                    "type": "string",
                    "example": "Qobulov"
                },
                "otp_code": {
                    "type": "string",
                    "example": "482910"
                },
                "password": {
                    "type": "string",
                    "example": "strong-password"
                },
                "phone": {
                    "type": "string",
                    "example": "+998901234567"
                },
                "username": {
                    "type": "string",
                    "example": "qobulov"
                }
            }
        },
        "authdto.RegisterResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/authdto.RegisterData"
                },
                "message": {
                    "type": "string",
                    "example": "Request processed successfully"
                },
                "meta": {
                    "$ref": "#/definitions/authdto.ResponseMeta"
                },
                "slug": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "authdto.ResetPasswordRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "new-strong-password"
                },
                "reset_token": {
                    "type": "string",
                    "example": "opaque-reset-token"
                }
            }
        },
...
        "authdto.ResetVerifyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/authdto.ResetVerifyData"
                },
                "message": {
                    "type": "string",
                    "example": "Request processed successfully"
                },
                "meta": {
                    "$ref": "#/definitions/authdto.ResponseMeta"
                },
                "slug": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "authdto.ResponseMeta": {
            "type": "object",
            "properties": {
                "api_version": {
                    "type": "string",
                    "example": "v1"
                },
                "duration": {
                    "type": "string",
                    "example": "2.22896ms"
                },
                "request_id": {
                    "type": "string",
                    "example": "aaddf89a9a081829a856613eff19682b"
                },
                "service": {
                    "type": "string",
                    "example": "brothers_app"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2026-09-22T12:55:03Z"
                }
            }
        },
        "authdto.SendOTPRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email is required for registration. For password_reset, provide exactly\none of Email or Username.",
                    "type": "string",
                    "example": "ali@example.com"
                },
                "purpose": {
                    "type": "string",
                    "enum": [
                        "registration",
                        "password_reset"
                    ],
                    "example": "registration"
                },
                "username": {
                    "type": "string",
                    "example": "qobulov"
                }
            }
        },
...
        "authdto.StartResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/authdto.StartData"
                },
                "message": {
                    "type": "string",
                    "example": "Request processed successfully"
                },
                "meta": {
                    "$ref": "#/definitions/authdto.ResponseMeta"
                },
                "slug": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "authdto.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "first_name": {
                    "type": "string",
                    "example": "Azizbek"
                },
                "language": {
                    "type": "string",
                    "enum": [
                        "uz",
                        "ru",
                        "en"
                    ],
                    "example": "uz"
                },
                "last_name": {
                    "type": "string",
                    "example": "Qobulov"
                }
            }
        },
...
        "authdto.UserResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/authdto.UserData"
                },
                "message": {
                    "type": "string",
                    "example": "Request processed successfully"
                },
                "meta": {
                    "$ref": "#/definitions/authdto.ResponseMeta"
                },
                "slug": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "responses.ErrorDetails": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string",
                    "example": "checking registration email: ERROR: column email does not exist (SQLSTATE 42703)"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Paste the access token only. Standard clients may send \"Bearer \u003ctoken\u003e\"; Swagger UI sends the token value directly.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}